hello

+++ Lab 2/lab2a/lab2a.c
@@ -138,8 +138,8 @@ int main (int argc, char **argv)
   /* print to STDOUT {elapsed_time, time_per_op} */
   //unsigned long long elapsed_time
   uint64_t elapsed_time = stop_time.tv_nsec - start_time.tv_nsec;
-  fprintf(stdout, "elapsed time: %llu ns\n", elapsed_time);
-  fprintf(stdout, "per operation: %llu ns\n", (elapsed_time/n_OPS));
+  fprintf(stdout, "elapsed time: %llu ns\n", (unsigned long long) elapsed_time);
+  fprintf(stdout, "per operation: %llu ns\n", (unsigned long long) (elapsed_time/n_OPS));
 
   /* Close the pthread_mutex if initiated */
   if(sync_type==SYNC_PTHREAD_MUTEX) pthread_mutex_destroy(&MUTEX_LOCK);
